+++ lib/inherited_state.dart
@@ -1,5 +1,5 @@
 library inherited_state;
 
-export 'src/inherited_state.dart' show InheritedState, InheritedSingleState, ReactiveState, RS, BuildContextExtension;
+export 'src/inherited_state.dart' show InheritedState, ReactiveState, RS, BuildContextEventExtension;
 export 'src/inject.dart' show Inject;
 export 'src/service_locator.dart' show ServiceLocator, SL, ServiceInject;